async_service/requirements.txt
fastapi==0.79.0
ja-sentence-segmenter==0.0.2
khmer-nltk==1.5
Morfessor==2.0.6
nltk==3.5
numpy==1.19.1
pika==1.3.0
Pillow==9.2.0
polyglot==16.7.4
pycld2==0.41
pydantic==1.9.1
PyICU==2.5
PyYAML==6.0
requests==2.28.1
sacremoses==0.0.53
sentencepiece==0.1.97
tqdm==4.64.0
transformers==4.18.0
uvicorn==0.18.2
pythainlp
ctranslate2==3.0.0